requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
